+++ CHANGELOG.md
@@ -6,4 +6,4 @@
 + Add keyword `const`.
 + Add keywords `if`, `then`, `else`.
 + Add handling for none-keyword data.
-
++ Add annotation keywords `examples`, `comment`